Main Button pad completed
Implemented the main numpad using the cursor position for dynamic editing.
Moved Logic to add to the expression string out of the viewModel and into a helper class to help keep the code modular and keep the viewmodel de-cluttered.
Added functions to the Viewmodel for the UI to access variable states in the helper class so the UI would not need to access/hold refences to the helper class